+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+v1.7.6
+- Fix capture logic
+- Fix compatibility with Magento 2.4.6
+
 v1.7.5
 - Added option allowing to configure which delivery method is pickup